Rename `RequireAuth` to `WithAuthentication`.
