+++ src/_components/WithAuthentication.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Navigate, useLocation } from 'react-router-dom';
 import { useAppSelector } from '../_state/hooks';
 
-export const RequireAuth: React.FC<{
+export const WithAuthentication: React.FC<{
   children: JSX.Element;
 }> = (props) => {
   const location = useLocation();
@@ -23,24 +23,3 @@ export const CheckAuth: React.FC<{ children: JSX.Element }> = (props) => {
   if (!user) return props.children;
   return <Navigate to={from} replace />;
 };
-
-// type RequireVerificationProps = { children: JSX.Element };
-// export const RequireVerification: React.FC<RequireVerificationProps> = (
-//   props
-// ) => {
-//   const location = useLocation();
-//   const navigate = useNavigate();
-//   const user = useAppSelector((state) => state.auth.user)!;
-//   const userMemo = useMemo(() => user, [user?.uid, user === null]);
-
-//   useEffect(() => {
-//     if (!userMemo.emailVerified || !userMemo.phoneNumber) {
-//       navigate('/verify', { state: { from: location } });
-//     }
-//   }, [userMemo]);
-
-//   // if (loading) return <LoadingPage />;
-
-//   // TODO: Handle verification state here.
-//   return props.children;
-// };
